+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased Changes
 
 - Added metadata output which includes scanned bytes (for cost tracking) to `execute_dql`
+- Added next-steps for `get_entity_details` to find out about metrics, problems and logs
 
 ## 0.5.0